Move NutRuntimeData to NUT coordinator module
Move `NutRuntimeData` dataclass from `__init__.py` into `coordinator.py`
alongside `NutConfigEntry` and `NutCoordinator`, and update all imports
in platform files accordingly.
