+++ homeassistant/components/nut/device_action.py
@@ -13,9 +13,8 @@ from homeassistant.core import Context, HomeAssistant
 from homeassistant.helpers import config_validation as cv, device_registry as dr
 from homeassistant.helpers.typing import ConfigType, TemplateVarsType
 
-from . import NutRuntimeData
 from .const import DOMAIN, INTEGRATION_SUPPORTED_COMMANDS
-from .coordinator import NutConfigEntry
+from .coordinator import NutConfigEntry, NutRuntimeData
 
 ACTION_TYPES = {cmd.replace(".", "_") for cmd in INTEGRATION_SUPPORTED_COMMANDS}
